Chat-with-csv--Using-Semantic-Search--main/requirements.txt
pandas>=1.5.0
streamlit>=1.28.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
llama-index>=0.9.0
llama-index-llms-ollama>=0.1.0
matplotlib>=3.7.0
numpy>=1.24.0